+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `tailspin` now reads lines in chunks and highlights them in parallel. This should improve performance especially when
   reading large files.
+- Changed `-t,--tail` flag to `-e,--start-at-end` to avoid confusion with `tail -f`
 
 ## 2.2.0